[book-tools] Add helper tools
To keep the history clean and in sync with the text
for easier book reading, I often have to rebase and
modify the whole history.

To keep everything working and to be able to restore
tags easily (rebasing removes them) I have a few helper
scripts. They have to be in an early commit to be able
to use them anywhere in history.