Bump Android Version Code
Every new release must increase the version code.
I'm uploading a new release after updating the
Hotwire package version. The version itself was
updated by modifying the original commit, to keep
the history clean for readers, but the versionCode
bump is not relevant for readers so I am putting
it in a separate new commit.